chore: moved the type declaration into the actions
